move command line utils to command line package
